Modification to remove update_grid_metadata CLI and incorporate this functionality within the standardise CLI.
